ai-content-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
notebooklm-py>=0.1.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
aiofiles>=23.0.0
cryptography>=42.0.0
